requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.1,<0.28
djangorestframework-api-key>=3.0.0,<3.1.0